Adjusted logic to avoid SOF
Love this project idea, thank you for going ahead and whipping this up!
Great way to avoid having to mutate the state array! but of course we get to this StackOverflow issue when we get towards the end of the array. ( - i realized after going ahead with my changes that maybe we just had the logic in the pickStudent control flow backwards ? ( should be without bang ? )) potential alternative fix for sure

Summary of changes:
I reworked it so that we mutate the array now so every pick hits a valid student

|Next steps ?|
|********|
| add a restart button |
| Dark mode |
| a pass option that shortcuts the logic allowing the student to be re-picked if the don't know the answer |
| Any other fun things ? |

Things I did:
  * redefined nameArray with state, initialized with same values from .env
  * added ' classSize' variable to statically check for displayToggle
  * adjusted logic of pickStudent() to now store pick as a num for index removal, set the name, make a soft copy of the state array via slice, splice off the picked name, and set the state to the new mutated array
  * removed logic check for pickStudent - check is happening in child component
These changes do make picked students array only useful for checking the number of picks we've done which could easily be refactored to be dependant on the newly shrinking students array - but i left it so it could be used in the future if we ever add extras